backend/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.7.0
attrs==24.2.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
distro==1.9.0
fastapi==0.115.6
frozenlist==1.5.0
greenlet==3.1.1
groq==0.13.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.11
langchain-core==0.3.25
langchain-groq==0.2.1
langchain-text-splitters==0.3.2
langsmith==0.2.3
multidict==6.1.0
numpy==2.2.0
openai==1.57.4
orjson==3.10.12
packaging==24.2
pillow==11.0.0
propcache==0.2.1
pydantic==2.10.3
pydantic_core==2.27.1
PyMySQL==1.1.1
pytesseract==0.3.13
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
tenacity==9.0.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.33.0
yarl==1.18.3
